requirements.txt
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
autoenv==1.0.0
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
coverage==4.5.4
coveralls==1.8.2
Django==2.2.5
django-rest-auth==0.9.5
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
docopt==0.6.2
httpie==1.0.3
idna==2.8
importlib-metadata==0.23
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.2
pep8==1.7.1
Pillow==6.1.0
# pkg-resources==0.0.0
pluggy==0.13.0
postgres==2.2.2
psycopg2==2.7.1
psycopg2-binary==2.8.3
py==1.8.0
pycodestyle==2.5.0
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.2
pytest==5.1.2
pytest-cov==2.7.1
pytest-django==3.5.1
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.5
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.6.0
